Remove HTTPClient dependency in hapi-fhir-base
