+++ hapi-fhir-base/src/main/java/ca/uhn/fhir/util/UrlUtil.java
@@ -31,9 +31,6 @@ import com.google.common.net.PercentEscaper;
 import jakarta.annotation.Nonnull;
 import jakarta.annotation.Nullable;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.http.NameValuePair;
-import org.apache.http.client.utils.URLEncodedUtils;
-import org.apache.http.message.BasicNameValuePair;
 import org.hl7.fhir.instance.model.api.IPrimitiveType;
 
 import java.io.UnsupportedEncodingException;
@@ -570,39 +567,6 @@ public class UrlUtil {
 			}
 		}
 		return theString;
-	}
-
-	public static List<NameValuePair> translateMatchUrl(String theMatchUrl) {
-		List<NameValuePair> parameters;
-		String matchUrl = theMatchUrl;
-		int questionMarkIndex = matchUrl.indexOf('?');
-		if (questionMarkIndex != -1) {
-			matchUrl = matchUrl.substring(questionMarkIndex + 1);
-		}
-
-		final String[] searchList = new String[] {"|", "=>=", "=<=", "=>", "=<"};
-		final String[] replacementList = new String[] {"%7C", "=%3E%3D", "=%3C%3D", "=%3E", "=%3C"};
-		matchUrl = StringUtils.replaceEach(matchUrl, searchList, replacementList);
-		if (matchUrl.contains(" ")) {
-			throw new InvalidRequestException(Msg.code(1744) + "Failed to parse match URL[" + theMatchUrl
-					+ "] - URL is invalid (must not contain spaces)");
-		}
-
-		parameters = URLEncodedUtils.parse((matchUrl), Constants.CHARSET_UTF8, '&');
-
-		// One issue that has happened before is people putting a "+" sign into an email address in a match URL
-		// and having that turn into a " ". Since spaces are never appropriate for email addresses, let's just
-		// assume they really meant "+".
-		for (int i = 0; i < parameters.size(); i++) {
-			NameValuePair next = parameters.get(i);
-			if (next.getName().equals("email") && next.getValue().contains(" ")) {
-				BasicNameValuePair newPair =
-						new BasicNameValuePair(next.getName(), next.getValue().replace(' ', '+'));
-				parameters.set(i, newPair);
-			}
-		}
-
-		return parameters;
 	}
 
 	/**
