+++ hapi-fhir-server/pom.xml
@@ -49,6 +49,11 @@
 			<artifactId>owasp-java-html-sanitizer</artifactId>
 		</dependency>
 
+		<dependency>
+			<groupId>org.apache.httpcomponents</groupId>
+			<artifactId>httpclient</artifactId>
+		</dependency>
+
 		<!-- 
 		Spring is added as an optional dependency just so that it
 		can be used for CORS
